Add "center-master" to address #32
This is a more complete overhaul of the previous commit to add a new feature
that allows the master view to remain "centered", even when no other
views are present.

In this iteration, the new startup argument (boolean) for enabling this behavior
is `center-master` . Omitting the arg will default the behavior to
`false`.

`center-master` will work in all layouts, including monocle.
When `center-master` is enabled:
  - in wide layout, the master view always occupies the same dimensions,
   regardless of how many other views are spawned.
  - in left,right layouts, when only the master view is spawned, the
    view is centered horizontally with the **width** set according to `ratio_master`.
  - in top,bottom layouts, when only the master view is spawned, the
    view is centered vertically with **height** set according to
    `ratio_master`. The intention here was to enable this behavior for
    vertical screen users.
  - in monocle layout, both height and width are set according to ratio_master.
  - in all layouts, outer_gaps are included in the width calculation
    (maybe they shouldn't be?)

----

a couple of test cases have been added for the arrange_wide and
arguments. All tests pass.

----

Open questions / still to implement:

- a command to toggle `center_master`
- determine what is the desired monocle behavior
- why, in 'wide' layouts, when there is no master view but there are
  before+after views does the ratio calculation use the complement of
  the `ratio_wide` (curious what the practical use case of this is)