re-add center_master for non-wide layouts + misc cleanups
I had originally added a block to arrange_master_stack to allow
centering the view even when the layout was not 'wide', but removed
it while testing something else, and forgot to add it back.  This adds that back.

Also some small cleanups to reduce the amount of diffs to upstream.
mostly whitespace fixes.